This function takes an unsigned integer b as argument, which represents the number of bytes to allocate. It then calls malloc to allocate memory of that size, and checks if the returned pointer is NULL (which indicates that the allocation failed). If the allocation failed, the function calls exit with a status value of 98, which causes the program to terminate. Otherwise, the function returns the allocated pointer.
Note that we're using the stdlib.h header file for malloc and exit. We're also not using any other C standard library functions besides malloc, free and exit, and we're not using any global variables. Finally, we're returning a void* pointer, which can be cast to any other pointer type as needed.