The function starts by checking if new_size is zero. If it is, then the function frees the memory pointed to by ptr and returns NULL.
Next, the function checks if ptr is NULL. If it is, then the function simply calls malloc with new_size and returns the new pointer.

If new_size is equal to old_size, then the function simply returns ptr.

If none of the above conditions are met, then the function allocates new memory with malloc with the size of new_size. If malloc returns NULL, then the function returns NULL.

Then, the function copies the contents of the old memory block pointed to by ptr to the new memory block pointed to by new_ptr. The copying is done up to the minimum of old_size and new_size.

Finally, the function frees the old memory block pointed to by ptr and returns the new memory block pointed to by new_ptr.